src/test/java/de/tum/in/ase/eist/QueryProcessorTest.java
package de.tum.in.ase.eist;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.fail;

class QueryProcessorTest {

	private final QueryProcessor queryProcessor = new QueryProcessor();

	@Test
	void testEmptyStringIfCannotProcessQuery() {
		assertEquals("", queryProcessor.process("test"));
	}

	@Test
	void testKnowsAboutShakespeare() {
		String actual = queryProcessor.process("Shakespeare");
		if (!actual.contains("playwright")) {
			fail("The QueryProcessor does not know about Shakespeare.");
		}
	}

	@Test
	void isNotCaseSensitive() {
		String actual = queryProcessor.process("shakespeare");
		if (!actual.contains("playwright")) {
			fail("Your QueryProcessor should not be case sensitive.");
		}
	}

	@Test
	void split() {
		assertEquals("12314213",queryProcessor.process("a085d130: which of the following numbers is the largest: 857, 123423, 134,434,12,12314213,34"));
		assertEquals("17",queryProcessor.process("what is 15 plus 2"));
	}

	@Test
	void m() {
		assertEquals("20",queryProcessor.process("what is 5 multiplied by 4"));

	}

}
